+++ packages/gallery/src/components/gallery/proGallery.js
@@ -4,13 +4,11 @@ import utils from '../../common/utils';
 import { viewModeWrapper } from '../../common/window/viewModeWrapper';
 import window from '../../common/window/windowWrapper';
 import { GalleryComponent } from '../galleryComponent';
+import '../../versionLogger';
 
 export default class ProGallery extends GalleryComponent {
   constructor(props) {
     super();
-    if (utils.isVerbose()) {
-      console.count('[OOISSR] proGallery constructor', window.isMock);
-    }
     const isSSR = !!window.isMock;
     this.canRender = !isSSR || props.allowSSR === true; //do not render if it is SSR
     if (this.canRender) {
@@ -26,9 +24,6 @@ export default class ProGallery extends GalleryComponent {
       viewModeWrapper.setViewMode(props.viewMode);
     }
     this.domId = props.domId || Math.floor(Math.random() * 10000);
-    if (utils.isVerbose()) {
-      console.log('[OOISSR] proGallery init', window.isMock);
-    }
   }
 
   componentWillReceiveProps(nextProps) {