+++ packages/gallery/src/versionLogger.js
@@ -1,0 +1,1 @@
+try {console.debug('Pro Gallery Version: 1.5.31'); } catch (e) {}